backend/requirements.txt
aiohttp==3.8.1
fastapi==0.95.1
Pillow==9.5.0
pydantic==1.10.6
uvicorn==0.20.0
openai==0.27.4